requirements.txt
streamlit==1.28.2
pandas==2.1.3
certifi==2024.2.2
databricks-sql-connector==3.0.3
azure-identity==1.15.0
azure-keyvault==4.2.0
azure-storage-blob==12.19.0